Improve StarStore migration to accommodate for half-migrations (#1710)

+++ src/tauon/t_modules/t_db_migrate.py
@@ -571,12 +571,17 @@ def database_migrate(
 		logging.info("Updating database to version 71")
 		logging.info("Creating a backup Star database star.p.bak71")
 		import pickle
-		with (user_directory / "star.p.bak71").open("wb") as file:
-			pickle.dump(tauon.star_store.db, file, protocol=pickle.HIGHEST_PROTOCOL)
+		backup_star_db = (user_directory / "star.p.bak71")
+		if not backup_star_db.exists():
+			with backup_star_db.open("wb") as file:
+				pickle.dump(tauon.star_store.db, file, protocol=pickle.HIGHEST_PROTOCOL)
 
 		new_starstore_db: dict[tuple[str, str, str], StarRecord] = {}
 		old_record: list[int | str] = [] # Here just for typing
 		for key, old_record in star_store.db.items():
+			if isinstance(old_record, StarRecord):
+				logging.warning("Record {old_record} was already a StarRecord, likely half-failed earlier migration, skipping this migration over…")
+				break
 			new_record = StarRecord()
 			new_record.playtime = old_record[0]
 			new_record.loved = "L" in old_record[1]
@@ -585,7 +590,7 @@ def database_migrate(
 			if len(old_record) == 4:
 				new_record.loved_timestamp = old_record[3]
 			new_starstore_db[key] = new_record
-
-		star_store.db = new_starstore_db
+		else:
+			star_store.db = new_starstore_db
 
 	return master_library, multi_playlist, star_store, p_force_queue, theme, prefs, gui, gen_codes, radio_playlists
